+++ src/database/config/database-config.type.ts
@@ -1,13 +1,11 @@
 export type DatabaseConfig = {
   isDocumentDatabase: boolean;
   url?: string;
-  type?: string;
   host?: string;
   port?: number;
   password?: string;
   name?: string;
   username?: string;
-  synchronize?: boolean;
   maxConnections: number;
   sslEnabled?: boolean;
   rejectUnauthorized?: boolean;