+++ src/database/seeds/relational/seed.module.ts
@@ -1,9 +1,6 @@
 import { Module } from '@nestjs/common';
 import { ConfigModule } from '@nestjs/config';
-import { TypeOrmModule } from '@nestjs/typeorm';
-
-import { DataSource, DataSourceOptions } from 'typeorm';
-import { TypeOrmConfigService } from '../../typeorm-config.service';
+import { PrismaModule } from '../../prisma.module';
 import { RoleSeedModule } from './role/role-seed.module';
 import { StatusSeedModule } from './status/status-seed.module';
 import { UserSeedModule } from './user/user-seed.module';
@@ -20,12 +17,7 @@ import appConfig from '../../../config/app.config';
       load: [databaseConfig, appConfig],
       envFilePath: ['.env'],
     }),
-    TypeOrmModule.forRootAsync({
-      useClass: TypeOrmConfigService,
-      dataSourceFactory: async (options: DataSourceOptions) => {
-        return new DataSource(options).initialize();
-      },
-    }),
+    PrismaModule,
   ],
 })
 export class SeedModule {}