+++ src/database/seeds/relational/status/status-seed.module.ts
@@ -1,10 +1,7 @@
 import { Module } from '@nestjs/common';
-import { TypeOrmModule } from '@nestjs/typeorm';
 import { StatusSeedService } from './status-seed.service';
-import { StatusEntity } from '../../../../statuses/infrastructure/persistence/relational/entities/status.entity';
 
 @Module({
-  imports: [TypeOrmModule.forFeature([StatusEntity])],
   providers: [StatusSeedService],
   exports: [StatusSeedService],
 })