+++ src/files/infrastructure/persistence/relational/mappers/file.mapper.ts
@@ -1,18 +1,20 @@
+import { Prisma, File as PrismaFile } from '@prisma/client';
 import { FileType } from '../../../../domain/file';
-import { FileEntity } from '../entities/file.entity';
 
 export class FileMapper {
-  static toDomain(raw: FileEntity): FileType {
+  static toDomain(raw: PrismaFile): FileType {
     const domainEntity = new FileType();
     domainEntity.id = raw.id;
     domainEntity.path = raw.path;
     return domainEntity;
   }
 
-  static toPersistence(domainEntity: FileType): FileEntity {
-    const persistenceEntity = new FileEntity();
-    persistenceEntity.id = domainEntity.id;
-    persistenceEntity.path = domainEntity.path;
-    return persistenceEntity;
+  static toPersistence(
+    domainEntity: FileType,
+  ): Prisma.FileUncheckedCreateInput {
+    return {
+      id: domainEntity.id,
+      path: domainEntity.path,
+    };
   }
 }