+++ src/session/infrastructure/persistence/relational/relational-persistence.module.ts
@@ -1,11 +1,10 @@
 import { Module } from '@nestjs/common';
 import { SessionRepository } from '../session.repository';
 import { SessionRelationalRepository } from './repositories/session.repository';
-import { TypeOrmModule } from '@nestjs/typeorm';
-import { SessionEntity } from './entities/session.entity';
+import { PrismaModule } from '../../../../database/prisma.module';
 
 @Module({
-  imports: [TypeOrmModule.forFeature([SessionEntity])],
+  imports: [PrismaModule],
   providers: [
     {
       provide: SessionRepository,