+++ src/users/infrastructure/persistence/relational/relational-persistence.module.ts
@@ -1,11 +1,10 @@
 import { Module } from '@nestjs/common';
 import { UserRepository } from '../user.repository';
 import { UsersRelationalRepository } from './repositories/user.repository';
-import { TypeOrmModule } from '@nestjs/typeorm';
-import { UserEntity } from './entities/user.entity';
+import { PrismaModule } from '../../../../database/prisma.module';
 
 @Module({
-  imports: [TypeOrmModule.forFeature([UserEntity])],
+  imports: [PrismaModule],
   providers: [
     {
       provide: UserRepository,